Generalize the hotkey label to all buttons as well.
Well, except the IconButton. IconButtons should probably display their
hotkeys in a tooltip or something.
